+++ hs-bindgen/src-internal/HsBindgen/Frontend/Analysis/IncludeGraph.hs
@@ -21,14 +21,14 @@ module HsBindgen.Frontend.Analysis.IncludeGraph (
   , getIncludes
     -- * Debugging
   , Predicate
+  , DumpOpts(..)
   , dumpMermaid
   ) where
 
-import Data.DynGraph.Labelled (DynGraph)
-import Data.DynGraph.Labelled qualified as DynGraph
+import Data.DynGraph.Labelled (DynGraph, MermaidOptions (reverseEdges))
+import Data.DynGraph.Labelled qualified as Graph
 import Data.List qualified as List
 import Data.Map.Strict qualified as Map
-import Data.Text qualified as Text
 
 import Clang.Paths
 
@@ -40,12 +40,14 @@ import HsBindgen.Imports
   Definition
 -------------------------------------------------------------------------------}
 
+type Graph = DynGraph
+
 -- | Include graph
 --
 -- We create a DAG of C header paths with an edge for each @#include@.
 -- The edges are /reversed/ to represent an \"included by\" relation.
 data IncludeGraph = IncludeGraph{
-      graph :: DynGraph Include SourcePath
+      graph :: Graph Include SourcePath
     }
   deriving stock (Show, Eq)
 
@@ -101,7 +103,7 @@ getIncludeMacroArg = \case
 -------------------------------------------------------------------------------}
 
 empty :: IncludeGraph
-empty = IncludeGraph DynGraph.empty
+empty = IncludeGraph Graph.empty
 
 register ::
      SourcePath -- ^ Path of header that includes the following header
@@ -110,7 +112,7 @@ register ::
   -> IncludeGraph
   -> IncludeGraph
 register header include incHeader includeGraph = IncludeGraph $
-    DynGraph.insertEdge incHeader include header includeGraph.graph
+    Graph.insertEdge incHeader include header includeGraph.graph
 
 fromList :: [(SourcePath, Include, SourcePath)] -> IncludeGraph
 fromList edges = List.foldl' add empty edges
@@ -123,11 +125,11 @@ fromList edges = List.foldl' add empty edges
 -------------------------------------------------------------------------------}
 
 reaches :: IncludeGraph -> SourcePath -> Set SourcePath
-reaches includeGraph = DynGraph.reaches includeGraph.graph . List.singleton
+reaches includeGraph = Graph.reaches includeGraph.graph . List.singleton
 
 toSortedList :: IncludeGraph -> [SourcePath]
 toSortedList includeGraph =
-    List.delete RootHeader.name $ DynGraph.topSort includeGraph.graph
+    List.delete RootHeader.name $ Graph.topSort includeGraph.graph
 
 toOrderMap :: IncludeGraph -> Map SourcePath Int
 toOrderMap graph = Map.fromList (zip (toSortedList graph) [0..])
@@ -135,34 +137,120 @@ toOrderMap graph = Map.fromList (zip (toSortedList graph) [0..])
 getIncludes ::
      IncludeGraph
   -> SourcePath
-  -> DynGraph.FindEdgesResult Include
-getIncludes includeGraph = DynGraph.findEdges includeGraph.graph
+  -> Graph.FindEdgesResult Include
+getIncludes includeGraph = Graph.findEdges includeGraph.graph
 
 {-------------------------------------------------------------------------------
   Debugging
 -------------------------------------------------------------------------------}
 
 -- | Include graph predicate
---
--- An include graph predicate can be used to filter which indexes of an include
--- graph are shown.
 type Predicate = SourcePath -> Bool
 
-dumpMermaid :: Predicate -> IncludeGraph -> String
-dumpMermaid p includeGraph =
-    DynGraph.dumpMermaid
-      DynGraph.MermaidOptions{
-          reverseEdges = True
-        , renderVertex = \path -> guard (p path) >> return (getSourcePath path)
-        , renderEdge   = Just . renderInclude
-        }
-      includeGraph.graph
+data DumpOpts = DumpOpts {
+      -- | Only show vertices satisfying the predicate.
+      --
+      --   Combine dangling (i.e., transient) edges at removed vertices.
+      --
+      --   For example,
+      --
+      --     A-D->B-->C
+      --         |
+      --         +-->D
+      --
+      --   Removal of node 'B' creates
+      --
+      --     A-->C
+      --     |
+      --     --->D
+      predicate :: Predicate
+      -- | How should we show the include header?
+      --
+      -- - If 'True': Show paths of include header files
+      --
+      -- - If 'False':  Show the @#include@ argument, which usually shorter
+    , showPaths :: Bool
+    }
+
+-- | Dump include graph
+--
+-- See 'DumpOpts'.
+--
+-- Render direct (transient) dependencies using straight (dotted) arrows,
+-- respectively.
+dumpMermaid :: DumpOpts -> IncludeGraph -> String
+dumpMermaid o g =
+    Graph.dumpMermaid opts $
+      Graph.combineParallelEdges combineParallel $
+        Graph.filterVerticesCombineEdges predicate combineSequential $
+          Graph.mapEdges (const Direct) $
+            Graph.mapVerticesOutgoingEdges Vertex g.graph
   where
-    renderInclude :: Include -> String
-    renderInclude = \case
-      BracketInclude     i   -> "#include <"       ++ i.path ++ ">"
-      QuoteInclude       i   -> "#include \""      ++ i.path ++ "\""
-      MacroInclude       _ m -> "#include "        ++ Text.unpack m
-      BracketIncludeNext i   -> "#include_next <"  ++ i.path ++ ">"
-      QuoteIncludeNext   i   -> "#include_next \"" ++ i.path ++ "\""
-      MacroIncludeNext   _ m -> "#include_next "   ++ Text.unpack m
+    predicate :: Vertex -> Bool
+    predicate v = o.predicate v.path
+
+    renderVertex :: Vertex -> FilePath
+    renderVertex v =
+      if o.showPaths then
+        getSourcePath v.path
+      else
+        getIncludePath v
+
+    renderEdge :: Edge -> Graph.EdgeSpec
+    renderEdge = \case
+      Direct    -> Graph.EdgeSpec Graph.Straight Nothing
+      Transient -> Graph.EdgeSpec Graph.Dotted   Nothing
+
+    opts :: Graph.MermaidOptions Edge Vertex
+    opts = Graph.MermaidOptions{
+        reverseEdges = True
+      , renderVertex = Just . renderVertex
+      , renderEdge   = renderEdge
+      }
+
+    toPath :: Include -> FilePath
+    toPath = (.path) . getIncludeArg
+
+    getIncludePath :: Vertex -> FilePath
+    getIncludePath v = safeHead $ List.sortOn length $  map toPath $ v.includes
+
+    safeHead :: [String] -> String
+    safeHead []    = getSourcePath $ RootHeader.name
+    safeHead (x:_) = x
+
+data Edge = Direct | Transient
+ deriving stock (Show, Eq, Ord)
+
+data Vertex = Vertex {
+      includes :: [Include]
+    , path     :: SourcePath
+   }
+ deriving stock (Show, Eq, Ord)
+
+-- | Sequential combination of simple include edges.
+--
+-- @
+-- A---D--->B---D--->C
+--
+-- A--------I------->C
+-- @
+combineSequential :: Edge -> Edge -> Edge
+combineSequential _ _ = Transient
+
+-- | Parallel combination of simple include edges.
+--
+-- @
+-- A---D--->B---D--->C
+-- |                 ^
+-- |                 |
+-- ---------D---------
+--
+-- A--------I------->C
+-- |                 ^
+-- |                 |
+-- ---------D---------
+-- @
+combineParallel :: Edge -> Edge -> Edge
+combineParallel (Direct   ) (_        ) = Direct
+combineParallel (_        ) (Direct   ) = Direct
+combineParallel (Transient) (Transient) = Transient